m3-ocr-and-translation/requirements.txt
Flask==2.3.2
gunicorn==20.1.0
requests==2.32.3
Pillow==10.4.0
openai==1.51.0
python-dotenv==1.0.0
